Move BiHashMap to crate::collections::bi_hash_map
As we add more things to the collections module, it will be important to
keep different implementations in different modules.

This re-exports the module so it can be used as it was before, directly
from crate::collections::.
